Extracted API info to data
