+++ src/data/bill/action.tsx
@@ -1,6 +1,6 @@
 import { API_URL } from "./const";
 
-export const action = async () => {
+export const fetchBillData = async () => {
   const data = await fetch(API_URL);
   const result = await data.json();
   console.log(result);
