+++ src/pages/Create.tsx
@@ -1,15 +1,14 @@
 import React from "react";
-import { IonContent, IonItem, IonPage, IonRouterLink } from "@ionic/react";
-import { ImageElem } from "../../common/image-elem";
-import ChevronLeft from "../../assets/svg/chevron-left.svg";
+import { IonContent, IonItem, IonRouterLink } from "@ionic/react";
+import { ImageElem } from "../common/image-elem";
+import ChevronLeft from "../assets/svg/chevron-left.svg";
 import styled from "styled-components";
-import { Toolbar } from "../../common/layout-toolbar";
-import { TextElem } from "../../common/text-elem";
-import { ItemWrapper } from "../../common/item-wrapper";
-import { ListLayout } from "../../common/list-layout";
-import { BillForm } from "../../epic/bill-create-form";
-import {StyledIonPage} from "../../theme/global.styled";
-
+import { Toolbar } from "../common/layout-toolbar";
+import { TextElem } from "../common/text-elem";
+import { ItemWrapper } from "../common/item-wrapper";
+import { ListLayout } from "../common/list-layout";
+import { BillForm } from "../epic/bill-create-form";
+import { StyledIonPage } from "../theme/global.styled";
 
 const BackButton = styled(IonItem)`
   --ion-item-background: #2c2c2c;
